requirements.txt
Flask==2.0.0
PyYAML==5.4.1
redis==3.5.3
